chore: address reviews
restructuring, using 'startswith', other updates
