+++ modmail/errors.py
@@ -8,3 +8,9 @@ class InvalidArgumentError(Exception):
     """Improper argument."""
 
     pass
+
+
+class ConfigLoadError(Exception):
+    """Exception if the configuration failed to load from a local file."""
+
+    pass
